Add tests for R-Type instructions (they all pass)
All R-Type instructions pass the tests with the exception of the SUB
instruction. This is a weird one because the verilator test doesn't
fail, which leads me to believe the problem lies in the scala backend.
I'm not 100% certain so I'll have to look into it with more attention.

Simplify right shift operand selection. We only need to look at the 10th
bit of the immediate. Also, fix some bugs in the R-Type operand
modifications, only change when the immediate, op code and funct3 all
match the rigth criteria.