Add cmp_flag to the ALU used only in Control instr
The validity of the cmp flag at the end of the ALU is dependent on the
state of the PC. When a control instruction is detected in the decoder
stage, it must pass the ALU the correct inputs. More over, when
executing a BEQ, BNE, BGE, BLT, BLTU, or a BGEU, the PC must interpret
the value of the cmp flag accordingly.

For example, when the decoder detects a BEQ instruction it should output
to the ALU a SUB op between rs1 and rs2, and inspect the cmp_flag. If
the flag is 0 then rs1 and rs2 have the same contents, which the PC
should interpret as a command to branch. In the case of a BLT
instruction, the decoder should output to the ALU a SLT instruction, and
inspect the cmp_flag. If the flag is 1 then rs1 is less than rs2, which
the PC should interpret as a command to branch.