Share interfaces among modules
Update the test benches accordingly

Instruction Decoder: has two output interfaces, one for the register
file (DecoderRegisterOut) and one for the ALU (DecoderALUOut). The ALU
interface contains the immediate, the ALU OP code and the instruction OP
code. The register file interface contains the selection signals for
the input register (rsd) and the output registers (rs1 and rs2).

Register File: has the one input interface shared with the instruction
decoder (DecoderRegisterOut), and one output interface shared with the
ALU (RegisterFileOut). The latter interface, consists of the contents of
the registers read.

ALU: Shares Interface with the output of the RegisterFile
(RegisterFileOut) and the decoder output for ALU parameters
(DecoderALUOut).