Add Instruction Fetch stage
Pc is a module that does the branch detection and calculates the next pc.
It receives 10 bit: 7 opcode bits + 3 function bits for conditional branch.
Base on the opcode it detects if the instruction decoded is a branch, then
goes to defining witch type of branch is it with the respective flag control.
It also receives a immediate value from decoder that is 32 bit extended used
as offset for the jump.
Flag for conditional branches is calculated in ALU.
Also receives a 'step' signal witch is the address to witch to jump for the
JALR instruction. It connects to ALU result since is the result of rd + imm
calculated in the ALU.
The 'pc_out' output connects to Instruction Memory.
Instruction Memory is constructed based on a vector and is asynchronous.
Don't know how is filled with data.
Both pc and isntrMem .scala files have constants classes that needs to be
relocated.