Add operand select control signals to the idecoder
Add operand_a select signal to choose between pc and rs1. The immediate
is always passed to the ALU to facilitate the arithmetic operations.
This may change down the line.

Add mux select for the write back stage.

Add another immediate which is only sent to the instruction fetch stage.
This is why it may be a bad design decision to add an immediate input to
the ALU...

Add a we for the memory

There is still no decoding for the FENCE, CSR and E instructions. These
will be added later down line. Currently, we are focusing on the basics.
This is bad code and I apologize. But right now I just to see if it
works.