Changes to InstrMem, idecode, Pc and Core
Logic to detect jumps in InstrMem was taken out and tranfered to decoder.
As result a "stall" output was added and as well a input to Pc to serve pause
pc. It also subtitutes the jump detection priviousely used.
Added the necessary connections in core.
Added logic to correct the error of instruction"0" being executed 2 times,
but not confident with it.
After this a jump error was detected to be corrected.