PC: Add tests for JALR instructions
This commit changes the interface of the PC module, specifically it
changes the type of the br_step input signal from UInt to SInt. This
modification was made because the unsignedness of the input was being
ignored when used inside this module.

Regardless, this is a temporary change as the JALR instruction is
incorrectly implemented. As of this commit the PC value stored is the
current one and not the future one (+4).