Add test framework
The testing framework works as follows:
Each program in Adept-TestFiles will be executed using both
simulation backends (verilator and FIRRTL), and then its output with
already verified corrcet outputs.
The script which will test all files is the run_instr_test.sh.

As an attempt to further modularise the core, a debug flag has been
added so that the user can obtain more details of the inworkings of
the CPU.