Fix bugs in InstructionControlSignals class
So turns out when calling an abstract method from the constructor of an
abstract class, requires that the class which implements said abstract
class already has its variables set. If there is a double loop, e.g., if
the dev expects the child class variables to be built before the
super class variables, it will generate a Null pointer exception.