Simplify ALU interface and decoder control signals
The ALU interface now only has three inputs: the op, operand A, and
operand B.

ALU operation is generated automatically via the getALUOp method. This
method receives the funct7, funct3 and opcode signals to translate the
function code into an ALU operation code.

The decoder gets two new outputs: switch_2_imm and immediate. The
switch_2_imm control signal selects the output of the MUX connected to
the ALU's operand B input. And the immediate is an input to the
previously described MUX.

Update the tests to reflect the interface changes.