Fixes SLLI, and adds tests
Change the file immediate.scala to check if the 7 MSBs follow the instruction
spec when the operation is a shift immediate. To test the correct execution,
the file AdeptTester.scala is also edited in order to print the correct values
of PC and registers when a trap is generated. In this file the first `step(1)`
is to set to low the signal load_we and the second `step(1)`, that is
introduced in this commit, is to assert reset. It's important to mention that,
when there is an assembly instruction that will generate a trap, said
instruction should be preceded by two NOPs so that the instructions in the
pipeline may be written to the register file.

This commit also adds the folder decoder in src/test/scala in order to create
chisel flat spec tests for this unit. Files decoderMain.scala,
decoderUnitTester.scala and imm_tests.scala implement already these tests
for SLLI instruction. In order to 'testOnly' modules with verilator the
Makefile was edited. The tests using FIRRTL are discarded by now because
there are some bugs with this testing backend.
The file AdeptUnitTester is also edited in order to print 'Trap' when
there is a trap.

Closes PR #8 and issue #2