Create ADDI Chisel flat spec tests for decoder
This commit adds the specific test for the instruction ADDI to the
chisel flat spec tests for the decoder. It's also change the
run_instr_tests script in order to parse the log resultant of
a test that generate a trap correctly, to have a match when
compare this parse with the expected result.